+++ scene/resources/skeleton_modification_stack_2d.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_2d.h                                           */
+/*  skeleton_modification_stack_2d.h                                     */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,58 +28,72 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef SKELETONMODIFICATION2D_H
-#define SKELETONMODIFICATION2D_H
+#ifndef SKELETONMODIFICATIONSTACK2D_H
+#define SKELETONMODIFICATIONSTACK2D_H
 
 #include "scene/2d/skeleton_2d.h"
-#include "scene/resources/skeleton_modification_stack_2d.h"
+#include "scene/resources/skeleton_modification_2d.h"
 
 ///////////////////////////////////////
-// SkeletonModification2D
+// SkeletonModificationStack2D
 ///////////////////////////////////////
 
-class SkeletonModificationStack2D;
+class Skeleton2D;
+class SkeletonModification2D;
 class Bone2D;
 
-class SkeletonModification2D : public Resource {
-	GDCLASS(SkeletonModification2D, Resource);
+class SkeletonModificationStack2D : public Resource {
+	GDCLASS(SkeletonModificationStack2D, Resource);
 	friend class Skeleton2D;
-	friend class Bone2D;
+	friend class SkeletonModification2D;
 
 protected:
 	static void _bind_methods();
-
-	SkeletonModificationStack2D *stack;
-	int execution_mode = 0; // 0 = process
-
-	bool enabled = true;
-	bool is_setup = false;
-
-	bool _print_execution_error(bool p_condition, String p_message);
+	void _get_property_list(List<PropertyInfo> *p_list) const;
+	bool _set(const StringName &p_path, const Variant &p_value);
+	bool _get(const StringName &p_path, Variant &r_ret) const;
 
 public:
-	virtual void _execute(float delta);
-	virtual void _setup_modification(SkeletonModificationStack2D *p_stack);
-	virtual void _draw_editor_gizmo();
+	Skeleton2D *skeleton = nullptr;
+	bool is_setup = false;
+	bool enabled = false;
+	float strength = 1.0;
 
-	bool editor_draw_gizmo = false;
-	void set_editor_draw_gizmo(bool p_draw_gizmo);
-	bool get_editor_draw_gizmo() const;
+	enum EXECUTION_MODE {
+		execution_mode_process,
+		execution_mode_physics_process
+	};
 
-	void set_enabled(bool p_enabled);
-	bool get_enabled();
+	Vector<Ref<SkeletonModification2D>> modifications = Vector<Ref<SkeletonModification2D>>();
 
-	Ref<SkeletonModificationStack2D> get_modification_stack();
-	void set_is_setup(bool p_setup);
+	void setup();
+	void execute(float delta, int p_execution_mode);
+
+	bool editor_gizmo_dirty = false;
+	void draw_editor_gizmos();
+	void set_editor_gizmos_dirty(bool p_dirty);
+
+	void enable_all_modifications(bool p_enable);
+	Ref<SkeletonModification2D> get_modification(int p_mod_idx) const;
+	void add_modification(Ref<SkeletonModification2D> p_mod);
+	void delete_modification(int p_mod_idx);
+	void set_modification(int p_mod_idx, Ref<SkeletonModification2D> p_mod);
+
+	void set_modification_count(int p_count);
+	int get_modification_count() const;
+
+	void set_skeleton(Skeleton2D *p_skeleton);
+	Skeleton2D *get_skeleton() const;
+
 	bool get_is_setup() const;
 
-	void set_execution_mode(int p_mode);
-	int get_execution_mode() const;
+	void set_enabled(bool p_enabled);
+	bool get_enabled() const;
 
-	float clamp_angle(float angle, float min_bound, float max_bound, bool invert_clamp = false);
-	void editor_draw_angle_constraints(Bone2D *operation_bone, float min_bound, float max_bound, bool constraint_enabled, bool constraint_in_localspace, bool constraint_inverted);
+	void set_strength(float p_strength);
+	float get_strength() const;
 
-	SkeletonModification2D();
+	SkeletonModificationStack2D();
 };
 
 #endif // SKELETONMODIFICATION2D_H
